Split sink- and source-specific code into different modules
