+++ modules/common/src/main/scala/com/snowplowanalytics/snowplow/postgres/config/LoaderConfig.scala
@@ -12,27 +12,13 @@
  */
 package com.snowplowanalytics.snowplow.postgres.config
 
-import java.util.{UUID, Date}
-import java.time.Instant
-
-import scala.jdk.CollectionConverters._
-
 import cats.syntax.either._
 
 import io.circe.Decoder
+import LoaderConfig.{JdbcUri, Purpose}
 import io.circe.generic.semiauto.deriveDecoder
-import io.circe.generic.extras.Configuration
-import io.circe.generic.extras.semiauto.deriveConfiguredDecoder
 
-import LoaderConfig.{JdbcUri, Source, Purpose}
-
-import software.amazon.awssdk.regions.Region
-import software.amazon.kinesis.common.InitialPositionInStream
-
-case class LoaderConfig(name: String,
-                        id: UUID,
-                        source: Source,
-                        host: String,
+case class LoaderConfig(host: String,
                         port: Int,
                         database: String,
                         username: String,
@@ -62,62 +48,6 @@ object LoaderConfig {
         case "JSON" => SelfDescribing.asRight
         case other => s"$other is not supported purpose, choose from ENRICHED_EVENTS and JSON".asLeft
       }
-  }
-
-  implicit val awsRegionDecoder: Decoder[Region] =
-    Decoder.decodeString.emap { s =>
-      val allRegions = Region.regions().asScala.toSet.map((r: Region) => r.id())
-      if (allRegions.contains(s)) Region.of(s).asRight
-      else s"Region $s is unknown, choose from [${allRegions.mkString(", ")}]".asLeft
-    }
-
-  sealed trait InitPosition {
-    /** Turn it into fs2-aws-compatible structure */
-    def unwrap: Either[InitialPositionInStream, Date] = this match {
-      case InitPosition.Latest => InitialPositionInStream.LATEST.asLeft
-      case InitPosition.TrimHorizon => InitialPositionInStream.TRIM_HORIZON.asLeft
-      case InitPosition.AtTimestamp(date) => Date.from(date).asRight
-    }
-  }
-  object InitPosition {
-    case object Latest extends InitPosition
-    case object TrimHorizon extends InitPosition
-    case class AtTimestamp(timestamp: Instant) extends InitPosition
-
-    implicit val ioCirceInitPositionDecoder: Decoder[InitPosition] =
-      Decoder.decodeJson.emap { json =>
-        json.asString match {
-          case Some("TRIM_HORIZON") => TrimHorizon.asRight
-          case Some("LATEST") => Latest.asRight
-          case Some(other) =>
-            s"Initial position $other is unknown. Choose from LATEST and TRIM_HORIZEON. AT_TIMESTAMP must provide the timestamp".asLeft
-          case None =>
-            val result = for {
-              root <- json.asObject.map(_.toMap)
-              atTimestamp <- root.get("AT_TIMESTAMP")
-              atTimestampObj <- atTimestamp.asObject.map(_.toMap)
-              timestampStr <- atTimestampObj.get("timestamp")
-              timestamp <- timestampStr.as[Instant].toOption
-            } yield AtTimestamp(timestamp)
-            result match {
-              case Some(atTimestamp) => atTimestamp.asRight
-              case None => "Initial position can be either LATEST or TRIM_HORIZON string or AT_TIMESTAMP object (e.g. 2020-06-03T00:00:00Z)".asLeft
-            }
-        }
-      }
-  }
-
-  sealed trait Source extends Product with Serializable
-  object Source {
-
-    case class Kinesis(appName: String, streamName: String, region: Region, initialPosition: InitPosition) extends Source
-    case class PubSub(projectId: String, subscriptionId: String) extends Source
-
-    implicit val config: Configuration =
-      Configuration.default.withSnakeCaseConstructorNames
-
-    implicit def ioCirceConfigSourceDecoder: Decoder[Source] =
-      deriveConfiguredDecoder[Source]
   }
 
   case class JdbcUri(host: String, port: Int, database: String, sslMode: String) {
